refactor:perceptor
