+++ src/anchors/PerceptorTab.tsx
@@ -1,8 +1,7 @@
 import $ from 'jquery';
 import { utils } from 'github-url-detection';
 import { isPerceptor, runsWhen, isPublicRepo } from '../utils/utils';
-import PerceptorBase from '../pages/ContentScripts/PerceptorBase';
-import { inject2Perceptor } from '../pages/ContentScripts/Perceptor';
+import PerceptorBase from './PerceptorBase';
 import { render } from 'react-dom';
 import React from 'react';
 import TeachingBubbleWrapper from '../views/TeachingBubbleWrapper';
