--- src/pages/ContentScripts/PerceptorBase.tsx
@@ -1,3 +1,0 @@
-export default abstract class PerceptorBase {
-  public abstract run(config: any | null): Promise<void>;
-}
